requirements.txt
python-dotenv==0.20.0
requests==2.32.3
py3-pinterest==0.1.0
mistralai==1.2.5